refactor: rename controller actions and remove unused
