+++ WebStoreAPI/Properties/launchSettings.json
@@ -11,7 +11,7 @@
   "profiles": {
     "WebStoreAPI": {
       "commandName": "Project",
-      "launchBrowser": false,
+      "launchBrowser": true,
       "launchUrl": "swagger",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
